mcp_tools/requirements.txt
fastapi
uvicorn
pandas
psycopg2-binary
openpyxl
meltano
python-multipart